refactor: extract health checks
